Search Bar - Adjust filter options and props
